+++ src/rules/selector-list-comma-space-after/index.js
@@ -1,4 +1,5 @@
 import {
+  isStandardRule,
   report,
   ruleMessages,
   styleSearch,
@@ -40,6 +41,7 @@ export default function (expectation) {
 
 export function selectorListCommaWhitespaceChecker({ locationChecker, root, result, checkedRuleName }) {
   root.walkRules(rule => {
+    if (!isStandardRule(rule)) { return }
     const selector = rule.selector
     styleSearch({ source: selector, target: ",", outsideFunctionalNotation: true }, match => {
       checkDelimiter(selector, match.startIndex, rule)